Split device folder gallery widgets
